requirements.txt
CairoSVG
Django
Unidecode
cssutils
django-ipware
django-jinja
django-sampledatahelper
django-sites
django-sr
djmail
easy-thumbnails
django-pglocks
gunicorn
jinja2
netaddr
premailer
psd-tools
psycopg2
python-dateutil
python-magic
pytz
requests
six
celery
django-celery-beat
django-celery-results
redis
django-notifications-hq
backoff
google-cloud
holidays
lxml
oauth2
simplejson
tqdm
tweepy
